ENH delegate runtime cleanup to resource_tracker
The resource tracker is now in charge of cleaning up shared
resources between between Python processes. It does so by
reference counting objects each time a Python process demands it
(this means the Python process sends a "REGISTER"/"UNREGISTER"
request to the resource_tracker). If an object's refcount
drops to 0, the resource tracker cleans it up. Previously,
Python processes themselves would clean up the shared objects,
and the resource_tracker would only intervene at the end of the program
when all other processes connected to it have exited.