Improve _kill_process_tree_without_psutil
Fix #356.
Follow-up #348.

Make the windows and the posix variants more symmetric.

For both platform families, if the system-specific tree kill method fails for
any reason other than process not found, log a warning message with the
traceback and recommand the user to install psutil, and try to terminate the
the root process to avoid the stalled shutdown observed in #356 on Windows.